+++ R/PLNnetwork.R
@@ -27,6 +27,7 @@
 ##'  \item{"ftol_out"}{outer solver stops when an optimization step changes the objective function by less than xtol multiply by the absolute value of the parameter. Default is 1e-6}
 ##'  \item{"maxit_out"}{outer solver stops when the number of iteration exceeds out.maxit. Default is 50}
 ##'  \item{"penalize_diagonal"}{boolean: should the diagonal terms be penalized in the graphical-Lasso? Default is FALSE.}
+##'  \item{"penalty_weights"}{p x p matrix of weights (defaut filled with 1) to adapt the amount of shrinkage to each pairs of node. Mus tbe symmetric.}
 ##' }
 ##'
 ##' The list of parameters \code{control_init} controls the optimization process in the initialization and in the function \code{\link[=PLN]{PLN}}, plus two additional parameters:
@@ -49,11 +50,12 @@ PLNnetwork <- function(formula, data, subset, weights, penalties = NULL, control
   args <- extract_model(match.call(expand.dots = FALSE), parent.frame())
 
   ## define default control parameters for optim and overwrite by user defined parameters
+  ctrl_main <- PLNnetwork_param(control_main, nrow(args$Y), ncol(args$Y),ncol(args$X), weighted = !missing(weights))
   if (is.null(control_init$trace)) control_init$trace <- 0
   ctrl_init <- PLN_param(control_init, nrow(args$Y), ncol(args$Y), ncol(args$X), weighted = !missing(weights))
   if (is.null(ctrl_init$nPenalties)) ctrl_init$nPenalties <- 30
   if (is.null(ctrl_init$min.ratio)) ctrl_init$min.ratio   <- .1
-  ctrl_main <- PLNnetwork_param(control_main, nrow(args$Y), ncol(args$Y),ncol(args$X), weighted = !missing(weights))
+  ctrl_init$penalty_weights <- ctrl_main$penalty_weights
 
   ## Instantiate the collection of models
   if (ctrl_main$trace > 0) cat("\n Initialization...")