+++ tests/testthat/test-plnnetworkfamily.R
@@ -15,9 +15,10 @@ test_that("PLNnetwork: main function, fields access and methods", {
   w <- rep(1, nrow(Y))
 
   ## extract the data matrices and weights
+  ctrl_main <- PLNmodels:::PLNnetwork_param(list(), nrow(Y), ncol(Y),ncol(X), weighted = FALSE)
   ctrl_init <- PLNmodels:::PLN_param(list(), nrow(Y), ncol(Y), ncol(X), weighted = FALSE)
   ctrl_init$trace <- 0; ctrl_init$nPenalties <- 30; ctrl_init$min.ratio   <- .1
-  ctrl_main <- PLNmodels:::PLNnetwork_param(list(), nrow(Y), ncol(Y),ncol(X), weighted = FALSE)
+  ctrl_init$penalty_weights <- ctrl_main$penalty_weights
 
   ## instantiate
   myPLN <- PLNmodels:::PLNnetworkfamily$new(NULL, Y, X, O, w, Abundance ~ 1, ctrl_init)
@@ -75,4 +76,36 @@ test_that("PLNnetwork computes the stability path only once.", {
                c(p, p))
   expect_length(extract_probs(models, index = 1, format = "vector"),
                p*(p-1)/2)
+})
+
+test_that("PLNnetwork: matrix of penalties work", {
+
+  p <- ncol(trichoptera$Abundance)
+  W <- diag(1, p, p)
+  W[upper.tri(W)] <- runif(p*(p-1)/2, min = 1, max = 5)
+  W[lower.tri(W)] <- t(W)[lower.tri(W)]
+  myPLN <- PLNnetwork(Abundance ~ 1, data = trichoptera, control_main = list(penalty_weights = W))
+
+  ## S3 methods
+  expect_true(PLNmodels:::isPLNnetworkfamily(myPLN))
+  expect_is(plot(myPLN), "ggplot")
+  expect_is(getBestModel(myPLN), "PLNnetworkfit")
+  expect_is(getModel(myPLN, myPLN$penalties[1]), "PLNnetworkfit")
+
+  ## Field access
+  expect_true(all(myPLN$penalties > 0))
+  expect_null(myPLN$stability_path)
+  expect_true(anyNA(myPLN$stability))
+
+  ## Other R6 methods
+  expect_true(is.data.frame(myPLN$coefficient_path()))
+  subs <- replicate(2,
+                    sample.int(nrow(trichoptera), size = nrow(trichoptera)/2),
+                    simplify = FALSE)
+  myPLN$stability_selection(subsamples = subs, control = list(penalty_weights = W))
+  expect_true(!is.null(myPLN$stability_path))
+  expect_true(inherits(myPLN$plot(), "ggplot"))
+  expect_true(inherits(myPLN$plot_objective(), "ggplot"))
+  expect_true(inherits(myPLN$plot_stars(), "ggplot"))
+
 })